// packages/common/src/di/interfaces/IProvider.ts
/**
 * @module common/di
 */
/** */
import {Type} from "@tsed/core";
import {ProviderType} from "./ProviderType";

/**
 *
 */
export interface IProvider<T> {
  /**
   * An injection token. (Typically an instance of `Type` or `InjectionToken`, but can be `any`).
   */
  provide: any;
  /**
   *
   */
  instance?: T;
  /**
   *
   */
  deps?: any[];
  /**
   * Class to instantiate for the `token`.
   */
  useClass?: Type<T>;
  /**
   *
   */
  useFactory?: Function;
  /**
   *
   */
  useValue?: any;

  /**
   * Provider type
   */
  type: ProviderType | any;

  /**
   *
   */
  [key: string]: any;
}